Finish work on final question
The final question asks about which species was observed the most at each park. The first code part the code seeks to find the answer and the second part of the code checks the simple max. It also includes some updates to the shape and info of each data set.